Calculator - Should return double when dividing two numbers (GREEN) - Quick implementation of double being returned by divide() in `calculator.java`

+++ src/main/java/utils/Calculator.java
@@ -14,7 +14,7 @@ public class Calculator {
         return a * b;
     }
 
-    public int divide(int a, int b) {
+    public double divide(double a, double b) {
         return a / b;
     }
